Added Support forum link
I quickly added a link to the support forum, didn't bother making it fancy or whatever, also I just quickly pasted in the Repl beta link there for testing purposes. The rest is up to you. If you'd like I could give the source code for the support forum so you can integrate it in Nebula.